Move dropdown closing logic to the global-navigation
