+++ libs/blocks/global-navigation/utilities/keyboard/utils.js
@@ -1,4 +1,7 @@
+import { selectors as baseSelectors } from '../utilities.js';
+
 const selectors = {
+  ...baseSelectors,
   mainNavItems:
     '.feds-navItem > a, .feds-navItem > .feds-cta-wrapper > .feds-cta',
   brand: '.feds-brand',
@@ -10,7 +13,6 @@ const selectors = {
   logo: '.gnav-logo',
   breadCrumbItems: '.feds-breadcrumbs li > a',
   expandedPopupTrigger: '.feds-navLink[aria-expanded = "true"]',
-  navLink: '.feds-navLink',
   promoLink: '.feds-promo-link',
   imagePromo: 'a.feds-promo-image',
   fedsNav: '.feds-nav',
@@ -19,9 +21,7 @@ const selectors = {
   section: '.feds-menu-section',
   column: '.feds-menu-column',
   cta: '.feds-cta',
-  curtain: '.feds-curtain',
   openSearch: '.feds-search-trigger[aria-expanded = "true"]',
-  globalNav: '.global-navigation',
 };
 
 selectors.popupItems = `
